requirements.txt
flask
flask-sqlalchemy
flask-cors
pyodbc
python-dotenv
